chore(openssh): Make it unlikely to hit `MaxStartups` rate limit by default
OpenSSH limits the number of unauthenticated connections to the SSH
daemon.

From the man page (`man 5 sshd_config`):

> **MaxStartups**
>
> Specifies the maximum number of concurrent unauthenticated connections
> to the SSH daemon.  Additional connections will be dropped until
> authentication succeeds or the **LoginGraceTime** expires for a connection.
> The default is 10:30:100.
>
> Alternatively, random early drop can be enabled by specifying the
> three colon separated values start:rate:full (e.g. "10:30:60").  sshd(8)
> will refuse connection attempts with a probability of rate/100 (30%) if
> there are currently start (10) unauthenticated connections.  The
> probability increases linearly and all connection attempts are refused
> if the number of unauthenticated connections reaches full (60).

Unfortunately it is not feasible to disable `MaxStartups` entirely,
because this value also determines how much memory is allocated for
the unauthenticated pipes.

Hence, we just increase it to a value that is large enough to make it
very unlikely to hit that limit.

If necessary, we can also consider to shorten the login grace time from
the very long default value of 120 seconds to something more reasonable,
like 30 seconds. That will allow the server to close unauthenticated
connections faster, which also decreases the likelyhood to run into the
rate limit.